pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "django-model-utils>=5.0.0",
    "djangorestframework>=3.17.1",
    "drf-yasg>=1.21.15",
    "python-dotenv>=1.2.2",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "ipython>=9.12.0",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.settings"
pythonpath = "src"
